refactor(app): extract discovery engine to internal/app (#32, step 2/3)
Moves the per-cycle helpers, credential probe code, loopConfig +
runDiscoveryLoop, and runCycle out of cmd/topology-exporter/main.go and
into the new internal/app package. main.go's run() now constructs an
app.LoopConfig and calls app.RunDiscoveryLoop.

New files:
  - internal/app/helpers.go     - DeduplicateOOS, MergeOOSFirstSeen,
                                  DeduplicateDevices, CollectDegradedReasons,
                                  SynthesizeEdges, ResolveEdgeDstDevices
  - internal/app/probe.go       - CredentialCandidate, CredentialCandidates,
                                  WalkSystemWithCredentials, ProfileToParams,
                                  Module, ModuleWalkFn
  - internal/app/cycle.go       - RunCycle, MaybeWarnLargeTopology,
                                  LargeTopologyEdgeThreshold,
                                  LargeTopologyWarnCooldownCycles
  - internal/app/loop.go        - LoopConfig, RunDiscoveryLoop,
                                  WarnSnapshot, OtlpPush,
                                  SnapshotWriteTimeout, OTLPPushTimeout,
                                  MaxOTLPPushConcurrency

main.go: 1427 -> 297 lines. The five-step refactor I originally planned
collapses to three commits in practice: steps 2-4 became this one move
to avoid export/unexport churn on helpers when runCycle landed alongside
them. One step remains - moving run() into internal/app.Run().

Symbols are exported within the internal/app package so cmd/topology-exporter
tests can keep referencing them via app.* prefix without test-file moves.
The `internal/` directory boundary already provides module-level encapsulation.

No behavior change. go build / go vet / go test -race ./... / gofmt all clean.
